v2

+++ prediction_module_ARIMA/prediction_module.py
@@ -6,7 +6,7 @@ from statsmodels.tsa.arima.model import ARIMA
 from kubernetes import client, config
 
 
-token = "FWe2VDkJXDmLp0o9Qid4"
+token = "NUuAendilxjox8GDVd03"
 org = "primary"
 bucket = "primary"
 
@@ -17,10 +17,11 @@ steps_ahead = 60
 config.load_incluster_config()
 v1 = client.AppsV1Api()
 
-th = 85
+base = 850
+th = 850
 
 while True:
-    with InfluxDBClient(url="http://influxdb:8086", token=token, org=org) as client:
+    with InfluxDBClient(url="http://influxdb.monitoring.svc:8086", token=token, org=org) as client:
         query = 'from(bucket: "primary") |> range(start: -10m) |> filter(fn: (r) => r["_measurement"] == "NetTraffic") |> filter(fn: (r) => r["Type"] == "Received") |> filter(fn: (r) => r["_field"] == "Size")'
         result = client.query_api().query(query, org=org)
 
@@ -31,16 +32,22 @@ while True:
                 points.append(record.get_value())
                 time_stamps.append(record.get_time())
 
+    if len(points)<50 : 
+        time.sleep(5)
+        continue
+
     
     model = ARIMA(points,order=(3,0,1))
     model_fit = model.fit()
     predicted_points = model_fit.forecast(steps=steps_ahead)
 
     last_time = time_stamps[-1]
-    with InfluxDBClient(url="http://influxdb:8086", token=token, org=org) as client:
+    with InfluxDBClient(url="http://influxdb.monitoring.svc:8086", token=token, org=org) as client:
         delete_api = client.delete_api()
         delete_api.delete(last_time+timedelta(seconds=3),last_time+timedelta(seconds=600), '_measurement="ARIMATrafficPrediction"', bucket='primary', org='primary')
         write_api = client.write_api(write_options=SYNCHRONOUS)
+        p = Point("Threshold").field("value", th)
+        write_api.write(bucket, org, p)
         for i in range(len(predicted_points)):
             p = Point("ARIMATrafficPrediction").tag("Algo", "ARIMA").field("Size", int(predicted_points[i])).time(last_time+timedelta(seconds=step_size*(i+1)))
             write_api.write(bucket, org, p) 
@@ -49,10 +56,11 @@ while True:
 
     if max(predicted_points)>th:
         print("current th :",th)
-        th = th*2 
+        ret = v1.read_namespaced_deployment_scale("helloweb","default")
+        cuurent_pods = ret.spec.replicas+1
+        th = base*cuurent_pods
         print("new th : ",th)
-        ret = v1.read_namespaced_deployment_scale("iperf-server-deployment","default")
         ret.spec.replicas = ret.spec.replicas+1
-        rpla = v1.replace_namespaced_deployment_scale("iperf-server-deployment","default", ret)
+        rpla = v1.replace_namespaced_deployment_scale("helloweb","default", ret)
         print(rpla)
     time.sleep(5)
